made new common button
